Update VERSION to be 'volatile'.
This prevents the compiler from replacing the field lookup with a constant.

This was causing the version number to not be updated in the About UI after it had been changed in the POM file. In this case, About.java was not changed, so it never got recompiled.

h/t @palbee for identifying the problem and for the suggested fix.